Added comments as discussed during last review
Currently, formalization.semantics contains HTML beautified expression (i.e. including <b> and <i> tags). They are currently removed in LTLSimLauncher (rewriteExpressionForLTLSIM) but this should probably be done before, such that the semantics object contains both, the plain and the HTML tagged expressions.

Currently the HTML tags are injected in SemanticsAnalyzer.replaceTemplateVarsWithArgs(), which has also a parameter noHTML to prevent injection of HTML, but it is always called with noHMTL = false in SemanticsAnalyzer.semantics().