Fixed deletion of text in template fields
Before: Deletion worked as expected when the cursor was at a single position inside a template field or a range of text inside a template field, excluding the leading and trailing blank characters, was selected. When a range including either the leading or the trailing blanks, removal of text did not work

After: Removal of text now works also when a range of text within a template field, including the leading or trailing blanks, is selected. However, when a range of text spanning outside the boundaries of the template field (i.e. including parts of the static text) is selected and the user hits delete or backspace, nothing is deleted.