Fixed parsing of text in editor when the selected template is changed
Before: When the selected template was changed, the text and fields in the editor were populated correctly. However, this new text was not immediately parsed (e.g. when the text contains errors, before the template fields are populated, for example when the default value for the timing field is "timing", those parse errors did not show). The text was only parsed once the user starts editoing, e.g. by placing the cursor into the editor

After: Now when the template selection changes the new text is parsed immediately. This is implemented by the `componentWilllReceiveProps()` method, which will check whether the new props correspond to a change of the template selection. If this is the case, parsing of the new editor text is triggered by `setContentInEditor()`